[REFACTOR] BackupType을 사용해 백업 전략 패턴 적용
- driverClass 기반 선택에서 BackupType(binlog, dump) 를 확인하여 전략 반환
