+++ src/main/java/ravo/ravobackend/coldStandbyBackup/backup/BackupStrategy.java
@@ -1,12 +1,13 @@
 package ravo.ravobackend.coldStandbyBackup.backup;
 
+import ravo.ravobackend.global.constants.BackupType;
 import ravo.ravobackend.global.domain.DatabaseProperties;
 
 import java.nio.file.Path;
 
 public interface BackupStrategy {
 
-    boolean support(String driverClassName);
+    boolean support(BackupType type);
 
     void backup(DatabaseProperties databaseProperties, Path backupDir) throws Exception;
 }
