+++ src/main/java/ravo/ravobackend/coldStandbyBackup/backup/BackupStrategyFactory.java
@@ -2,6 +2,7 @@ package ravo.ravobackend.coldStandbyBackup.backup;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
+import ravo.ravobackend.global.constants.BackupType;
 
 import java.util.List;
 
@@ -11,10 +12,10 @@ public class BackupStrategyFactory {
 
     private final List<BackupStrategy> strategies;
 
-    public BackupStrategy getBackupStrategy(String driverClassName) {
+    public BackupStrategy getBackupStrategy(BackupType type) {
         return strategies.stream()
-                .filter(strategy -> strategy.support(driverClassName))
+                .filter(strategy -> strategy.support(type))
                 .findFirst()
-                .orElseThrow(() -> new IllegalArgumentException("지원하는 백업 전략이 없습니다 : " + driverClassName));
+                .orElseThrow(() -> new IllegalArgumentException("지원하는 백업 전략이 없습니다 : " + type));
     }
 }
