+++ src/main/java/ravo/ravobackend/coldStandbyBackup/backup/dump/MySqlBackupStrategy.java
@@ -1,8 +1,10 @@
-package ravo.ravobackend.coldStandbyBackup.backup;
+package ravo.ravobackend.coldStandbyBackup.backup.dump;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
+import ravo.ravobackend.coldStandbyBackup.backup.BackupStrategy;
+import ravo.ravobackend.global.constants.BackupType;
 import ravo.ravobackend.global.domain.DatabaseProperties;
 import ravo.ravobackend.global.util.CommandRequest;
 import ravo.ravobackend.global.util.ShellCommandExecutor;
@@ -21,8 +23,8 @@ public class MySqlBackupStrategy implements BackupStrategy {
     private final ShellCommandExecutor shellCommandExecutor;
 
     @Override
-    public boolean support(String driverClassName) {
-        return driverClassName.toLowerCase().contains("mysql");
+    public boolean support(BackupType type) {
+        return type.equals(BackupType.MYSQL_DUMP);
     }
 
     @Override
