+++ src/main/java/ravo/ravobackend/global/constants/BackupType.java
@@ -1,0 +1,6 @@
+package ravo.ravobackend.global.constants;
+
+public enum BackupType {
+    MYSQL_BINLOG,
+    MYSQL_DUMP,
+}
