+++ src/test/java/ravo/ravobackend/coldStandbyBackup/backup/MySqlBackupStrategyTest.java
@@ -10,6 +10,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
+import ravo.ravobackend.coldStandbyBackup.backup.dump.MySqlBackupStrategy;
 import ravo.ravobackend.global.domain.DatabaseProperties;
 
 import java.io.File;
